refactor: default blocks data
