+++ apps/web/app/composables/useBlockTemplates/index.ts
@@ -1,6 +1,6 @@
 export * from './useBlockTemplates';
 export * from './types';
-export * from './homepage';
-export * from './category';
-export * from './product';
-export * from './templateLoader';
+export * from '../../utils/blockTemplates/homepage';
+export * from '../../utils/blockTemplates/category';
+export * from '../../utils/blockTemplates/product';
+export * from '../../utils/blockTemplates/templateLoader';
