+++ apps/web/app/utils/blockTemplates/category/interface.ts
