+++ apps/web/app/utils/blockTemplates/homepage/-.de.ts
