+++ apps/web/app/utils/blockTemplates/homepage/-.en.ts
