+++ apps/web/app/utils/blockTemplates/homepage/factory.ts
