+++ apps/web/app/utils/blockTemplates/homepage/index.ts
