+++ apps/web/app/utils/blockTemplates/homepage/interface.ts
