+++ apps/web/app/utils/blockTemplates/product/factory.ts
