+++ apps/web/app/utils/blockTemplates/product/index.ts
