+++ apps/web/app/utils/blockTemplates/product/interface.ts
